Second part of mtu= per conn commit

+++ programs/pluto/connections.h
@@ -266,7 +266,7 @@ struct connection {
     char *cisco_banner;
 #endif /* XAUTH */
     u_int8_t metric;              /* metric for tunnel routes */
-    u_int8_t connmtu;              /* mtu for tunnel routes */
+    u_int16_t connmtu;              /* mtu for tunnel routes */
 #ifdef HAVE_STATSD
     u_int32_t statsval;			/* track what we have told statsd */
 #endif
